# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/messages": {
      "get": {
        "operationId": "AdminMessagesController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "unread",
                "read",
                "archived"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminMessages"
        ]
      }
    },
    "/admin/messages/{id}": {
      "get": {
        "operationId": "AdminMessagesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminMessages"
        ]
      },
      "patch": {
        "operationId": "AdminMessagesController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMessageStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminMessages"
        ]
      },
      "delete": {
        "operationId": "AdminMessagesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminMessages"
        ]
      }
    },
...
    "/portal/documents/{id}/file": {
      "get": {
        "operationId": "PortalDocumentsController_streamFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PortalDocuments"
        ]
      }
    },
...
    "/admin/applications": {
      "get": {
        "operationId": "AdminApplicationsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reviewerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/counts": {
      "get": {
        "operationId": "AdminApplicationsController_counts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/assignees": {
      "get": {
        "operationId": "AdminApplicationsController_assignees",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/export.csv": {
      "get": {
        "operationId": "AdminApplicationsController_exportCsv",
        "parameters": [
          {
            "name": "reviewerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/bulk": {
      "post": {
        "operationId": "AdminApplicationsController_bulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/{id}": {
      "get": {
        "operationId": "AdminApplicationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      },
      "patch": {
        "operationId": "AdminApplicationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/{id}/request-documents": {
      "post": {
        "operationId": "AdminApplicationsController_requestDocuments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestDocumentsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/{id}/documents/{docId}": {
      "patch": {
        "operationId": "AdminApplicationsController_reviewDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/{id}/documents/{docId}/file": {
      "get": {
        "operationId": "AdminApplicationsController_streamDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
    "/admin/applications/{id}/notes": {
      "post": {
        "operationId": "AdminApplicationsController_addNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminApplications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Safeer API",
    "description": "جمعية سفير الدعوية / Safeer Association — content and admin API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sf_sid"
      }
    },
    "schemas": {
...
      "SetMessageStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "unread",
              "read",
              "archived"
            ]
          }
        },
        "required": [
          "status"
        ],
        "additionalProperties": false
      },
...
      "BulkActionDto": {
        "type": "object",
        "properties": {
          "ids": {
            "minItems": 1,
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^\\d+$"
            }
          },
          "action": {
            "type": "string",
            "enum": [
              "assign",
              "status",
              "request_documents"
            ]
          },
          "reviewerId": {
            "type": "string",
            "pattern": "^\\d+$",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "new",
              "under_review",
              "docs_missing",
              "interview",
              "accepted",
              "rejected"
            ]
          },
          "docTypes": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "id_copy",
                "certificate",
                "admission_letter",
                "other"
              ]
            }
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          }
        },
        "required": [
          "ids",
          "action"
        ],
        "additionalProperties": false
      },
      "UpdateApplicationAdminDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "new",
              "under_review",
              "docs_missing",
              "interview",
              "accepted",
              "rejected"
            ]
          },
          "assignedReviewerId": {
            "type": "string",
            "pattern": "^\\d+$",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RequestDocumentsDto": {
        "type": "object",
        "properties": {
          "docTypes": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "id_copy",
                "certificate",
                "admission_letter",
                "other"
              ]
            }
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          }
        },
        "required": [
          "docTypes"
        ],
        "additionalProperties": false
      },
      "ReviewDocumentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "rejected"
            ]
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "required": [
          "status"
        ],
        "additionalProperties": false
      },
      "CreateApplicationNoteDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          }
        },
        "required": [
          "body"
        ],
        "additionalProperties": false
      },
...
    }
  }
}